move report service test to integrationTest
