+++ rundeckapp/src/test/groovy/rundeck/services/ReportServiceSpec.groovy
@@ -206,71 +206,6 @@ class ReportServiceSpec extends Specification implements ServiceUnitTest<ReportS
         false   | Query.In
     }
 
-    def "report should include referenced executions"() {
-        given:
-        String jobUuid = UUID.randomUUID().toString()
-        service.referencedExecutionDataProvider = new GormReferencedExecutionDataProvider()
-        service.execReportDataProvider = Mock(ExecReportDataProvider)
-
-        def jobname = 'abc'
-        def group = 'path'
-        def project = 'AProject'
-        ScheduledExecution job = new ScheduledExecution(
-                uuid: jobUuid,
-                jobName: jobname,
-                project: project,
-                groupPath: group,
-                description: 'a job',
-                argString: '-args b -args2 d',
-                workflow: new Workflow(
-                        keepgoing: true,
-                        commands: [new CommandExec(
-                                [adhocRemoteString: 'test buddy', argString: '-delay 12 -monkey cheese -particle']
-                        )]
-                ),
-                retry: '1'
-        )
-        job.save()
-        ScheduledExecution job2 = new ScheduledExecution(
-                uuid: UUID.randomUUID().toString(),
-                jobName: jobname,
-                project: project,
-                groupPath: group,
-                description: 'a job',
-                argString: '-args b -args2 d',
-                workflow: new Workflow(
-                        keepgoing: true,
-                        commands: [new CommandExec(
-                                [adhocRemoteString: 'test buddy', argString: '-delay 12 -monkey cheese -particle']
-                        )]
-                ),
-                retry: '1'
-        )
-        job2.save()
-        Execution e1 = new Execution(
-                project: project,
-                scheduledExecution: job2,
-                user: 'bob',
-                dateStarted: new Date(),
-                dateEnded: new Date(),
-                status: 'successful'
-
-        )
-        e1.save()
-        ReferencedExecution refexec = new ReferencedExecution(status: 'success', jobUuid: job.uuid, execution: e1)
-        refexec.save()
-        ExecQuery query = new ExecQuery()
-        query.projFilter = "AProject"
-        query.jobIdFilter = "${job.id}"
-        query.execProjects = ["AProject"]
-        query.excludeJobListFilter = []
-        when:
-        def result = service.getExecutionReports(query, true)
-        then:
-        result.reports.size() == 1
-        1 * service.execReportDataProvider.getExecutionReports(_, _, _, [e1.id]) >> [new ExecReport(executionId: e1.id, dateCompleted: new Date())]
-    }
-
 
     def "delete by execution with uuid"() {
         given:
